requirements.txt
numpy>=1.19.4
pygame>=2.0.1
pyyaml>=5.3.1
